refactor: Update font styles and add Google Font in app.css and app.html
The font styles in app.css and app.html have been updated to use the "IBM Plex Mono" font family with a font weight of 600. Additionally, a link to the Google Font API has been added to app.html to load the font. This refactor improves the consistency and aesthetics of the application.